crm/bridge/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
psycopg2-binary==2.9.11
python-dotenv==1.0.1
